back-5/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
python-dotenv==1.0.0
boto3==1.29.3
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
websockets==12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
aiofiles==23.2.1
python-magic==0.4.27
email-validator==2.1.0.post1
python-magic-bin==0.4.14; sys_platform == 'win32'
PyJWT==2.8.0 